+++ mandatory/src/print/print_p.c
@@ -6,7 +6,7 @@
 /*   By: xvoorvaa <xvoorvaa@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/06/07 16:04:30 by xvoorvaa      #+#    #+#                 */
-/*   Updated: 2021/07/12 13:17:58 by xander        ########   odam.nl         */
+/*   Updated: 2021/09/02 12:06:40 by xander        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@
 
 int	print_p(va_list ap)
 {
-	int					len;
-	unsigned long long	hex;
+	int				len;
+	unsigned long	hex;
 
 	len = 0;
 	hex = (unsigned long long) va_arg(ap, void *);
